gui-e2e: interface test-code with framework.
Within 'tests/gui_e2e/test_piggyback_hub_conf.py'

+ interface the method with test-framework's CLEANUP environment.
+ add logging messages.

+++ tests/gui_e2e/test_piggyback_hub_conf.py
@@ -4,10 +4,11 @@
 # conditions defined in the file COPYING, which is part of this source code package.
 import enum
 import logging
+import os
 from collections.abc import Iterator, Mapping, Sequence
 from contextlib import contextmanager
 from pathlib import Path
-from typing import assert_never
+from typing import assert_never, Final
 
 import pytest
 
@@ -22,6 +23,7 @@ from tests.testlib.common.utils import wait_until
 from tests.testlib.site import Site
 
 logger = logging.getLogger(__name__)
+CLEANUP: Final[bool] = os.getenv("CLEANUP", "1") == "1"
 
 
 SITE_SPECIFIC_SETTINGS_REL_PATH = Path("etc/check_mk/multisite.d/sites.mk")
@@ -140,7 +142,7 @@ def _back_up_original_site_file_states(
 
     all_sites = [central_site] + remote_sites
     setting_files = (
-        [GLOBAL_SETTINGS_REL_PATH, SITE_SPECIFIC_SETTINGS_REL_PATH, Path("etc/omd/site.conf")]
+        [GLOBAL_SETTINGS_REL_PATH, SITE_SPECIFIC_SETTINGS_REL_PATH, SITE_CONF_REL_PATH]
         + [replication_changes_rel_path(site.id) for site in all_sites]
         + [replication_status_rel_path(site.id) for site in all_sites]
     )
@@ -158,11 +160,16 @@ def _setup_settings(
     site_specific_settings: Mapping[str, Mapping[str, object]] | None,
     central_site: Site,
     remote_sites: list[Site],
+    cleanup: bool = True,
 ) -> Iterator[None]:
-    """Backup all relevant settings, apply global and site-specific settings that need to be set up
-    as precondition to the tests, then restore original settings after the test"""
-    setting_files, backed_settings = _back_up_original_site_file_states(central_site, remote_sites)
+    """Backup all relevant site-specific settings.
 
+    Apply global and site-specific settings that need to be set up as precondition to the tests,
+    then restore original settings after the test.
+    """
+    setting_files, backed_settings = _back_up_original_site_file_states(central_site, remote_sites)
+    list_of_files = ", ".join(map(str, setting_files))
+    logger.info("Backup settings within: '%s'", list_of_files)
     if global_settings:
         central_site.update_global_settings(GLOBAL_SETTINGS_REL_PATH, dict(global_settings))
     if site_specific_settings:
@@ -176,11 +183,13 @@ def _setup_settings(
     try:
         yield
     finally:
-        for path in setting_files:
-            if path in backed_settings:
-                central_site.write_file(path, backed_settings[path])
-            elif central_site.file_exists(path):
-                central_site.delete_file(path)
+        if cleanup:
+            logger.info("Restore settings within: '%s'", list_of_files)
+            for path in setting_files:
+                if path in backed_settings:
+                    central_site.write_file(path, backed_settings[path])
+                elif central_site.file_exists(path):
+                    central_site.delete_file(path)
 
 
 def _wait_for_file_change(site: Site, file_path: Path, original_mtime: float) -> None:
@@ -223,7 +232,9 @@ def test_disabled_on_central__enable_on_remote__error(
 ) -> None:
     """Test that enabling the piggyback-hub site-specific for a remote site fails if it is not enabled for the central site"""
     # given
-    with _setup_settings(global_settings, site_specific_settings, test_site, [remote_site]):
+    with _setup_settings(
+        global_settings, site_specific_settings, test_site, [remote_site], CLEANUP
+    ):
         original_settings = test_site.read_site_specific_settings(SITE_SPECIFIC_SETTINGS_REL_PATH)[
             "sites"
         ][remote_site.id]
@@ -277,7 +288,9 @@ def test_enabled_on_central__enable_on_remote__no_error(
     enable_action: HubEnableActions,
 ) -> None:
     """Test that enabling the piggyback-hub site-specific for a remote site works if it is enabled for the central site"""
-    with _setup_settings(global_settings, site_specific_settings, test_site, [remote_site]):
+    with _setup_settings(
+        global_settings, site_specific_settings, test_site, [remote_site], CLEANUP
+    ):
         original_mtime = test_site.file_mtime(SITE_SPECIFIC_SETTINGS_REL_PATH)
 
         # when
@@ -330,7 +343,9 @@ def test_enabled_on_remote__disable_on_remote__no_error(
 ) -> None:
     """Test that disabling the piggyback-hub site-specific for a remote site works in general"""
     # given
-    with _setup_settings(global_settings, site_specific_settings, test_site, [remote_site]):
+    with _setup_settings(
+        global_settings, site_specific_settings, test_site, [remote_site], CLEANUP
+    ):
         original_mtime = test_site.file_mtime(SITE_SPECIFIC_SETTINGS_REL_PATH)
 
         # when
@@ -381,7 +396,9 @@ def test_enabled_on_remote__disable_on_central__error(
     disable_action: HubDisableActions,
 ) -> None:
     """Test that disabling the piggyback-hub site-specific for the central site fails if it is enabled for a remote site"""
-    with _setup_settings(global_settings, site_specific_settings, test_site, [remote_site]):
+    with _setup_settings(
+        global_settings, site_specific_settings, test_site, [remote_site], CLEANUP
+    ):
         original_settings = (
             test_site.read_site_specific_settings(SITE_SPECIFIC_SETTINGS_REL_PATH)["sites"][
                 remote_site.id
@@ -424,6 +441,7 @@ def test_enabled_on_remote__disable_on_central_by_reset__error(
         },
         test_site,
         [remote_site],
+        CLEANUP,
     ):
         original_settings = test_site.read_site_specific_settings(SITE_SPECIFIC_SETTINGS_REL_PATH)[
             "sites"
@@ -488,7 +506,9 @@ def test_disabled_on_remote_site__disable_on_central__no_error(
 ) -> None:
     """Test that disabling the piggyback-hub site-specific for the central site works if it is not enabled for any remote sites"""
     # given
-    with _setup_settings(global_settings, site_specific_settings, test_site, [remote_site]):
+    with _setup_settings(
+        global_settings, site_specific_settings, test_site, [remote_site], CLEANUP
+    ):
         original_mtime = test_site.file_mtime(SITE_SPECIFIC_SETTINGS_REL_PATH)
 
         # when
@@ -526,6 +546,7 @@ def test_enabled_on_remote__disable_globally__error(
         {"gui_e2e_remote": {"site_piggyback_hub": True}},
         test_site,
         [remote_site],
+        CLEANUP,
     ):
         original_settings = test_site.read_global_settings(GLOBAL_SETTINGS_REL_PATH)
 
@@ -575,7 +596,9 @@ def test_disabled_on_remote_or_enabled_on_central__disable_globally__no_error(
 ) -> None:
     """Test that disabling the piggyback-hub globally works if it is not enabled for any remote sites or it is enabled for the central site"""
     # given
-    with _setup_settings(global_settings, site_specific_settings, test_site, [remote_site]):
+    with _setup_settings(
+        global_settings, site_specific_settings, test_site, [remote_site], CLEANUP
+    ):
         original_mtime = test_site.file_mtime(GLOBAL_SETTINGS_REL_PATH)
 
         # when
@@ -609,6 +632,7 @@ def test_disabled_on_central__enable_globally__error(
         {"gui_e2e_central": {"site_piggyback_hub": False}},
         test_site,
         [remote_site],
+        CLEANUP,
     ):
         original_settings = test_site.read_global_settings(GLOBAL_SETTINGS_REL_PATH)
 
@@ -639,7 +663,7 @@ def test_unset_on_central_and_remote__enable_globally__no_error(
 ) -> None:
     """Test that enabling the piggyback-hub globally works if it is not disabled for the central site"""
     # given
-    with _setup_settings(None, None, test_site, [remote_site]):
+    with _setup_settings(None, None, test_site, [remote_site], CLEANUP):
         original_mtime = test_site.file_mtime(GLOBAL_SETTINGS_REL_PATH)
 
         # when
